cpu was parsed as a float instead of an int

+++ src/main.rs
@@ -71,7 +71,7 @@ fn main() {
 fn get_cpu_temp() -> f64 {
     fs::read_to_string("/sys/class/thermal/thermal_zone0/temp")
         .expect("Could not read cpu temperature from sysfs")
-        .parse::<f64>()
-        .expect("Could not parse CPU temp from sysfs")
+        .parse::<u64>()
+        .expect("Could not parse CPU temp from sysfs") as f64
         / 1000.
 }
